+++ lib/widgets/misc/desktop_menu.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
@@ -8,6 +6,7 @@ import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/create/create_entry.dart';
 import 'package:lotti/services/nav_service.dart';
 import 'package:lotti/utils/consts.dart';
+import 'package:lotti/utils/platform.dart';
 
 class DesktopMenuWrapper extends StatelessWidget {
   DesktopMenuWrapper({
@@ -20,7 +19,7 @@ class DesktopMenuWrapper extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    if (!Platform.isMacOS) {
+    if (!isDesktop) {
       return child;
     }